Refactored & fixed tests.
